refactor(gateway): migrate Signal adapter to bundled plugin
Move the Signal adapter and its rate-limit companion from
gateway/platforms/signal*.py into a self-contained bundled plugin under
plugins/platforms/signal/, following the Discord/Mattermost/Home
Assistant/Yuanbao migration shape.

register() supplies the hooks that previously lived as per-platform wiring
in core:
  - adapter_factory      -> the elif in gateway/run.py::_create_adapter()
  - check_fn             -> check_signal_requirements guard there
  - is_connected         -> the Platform.SIGNAL lambda in gateway/config.py
  - setup_fn             -> _setup_signal + _PLATFORMS entry + _builtin_setup_fn
  - apply_yaml_config_fn -> the signal_cfg block in load_gateway_config()
  - standalone_sender_fn -> _send_signal in tools/send_message_tool.py
  - cron_deliver_env_var -> SIGNAL_HOME_CHANNEL

The full signal-cli JSON-RPC sender (attachment batching + the shared
SignalAttachmentScheduler rate-limit state) moved into the plugin as
_standalone_send; tools/send_message_tool.py keeps a thin _send_signal shim
that adapts the legacy (extra, ...) signature so the in-module media + chunk
dispatch paths are unchanged. Signal is HTTP to a signal-cli daemon, so unlike
WebSocket-singleton platforms cron delivery works out-of-process.

Deliberately left generic in core: the Platform.SIGNAL enum literal, the
_apply_env_overrides SIGNAL_* bridge, and the _is_user_authorized allowlist
maps (SIGNAL_ALLOWED_USERS + SIGNAL_GROUP_ALLOWED_USERS stay there).

All consumer imports rewritten to plugins.platforms.signal.{adapter,
signal_rate_limit}. Updated the connection-checker guard test to exclude
bundled-plugin platforms whose is_connected lives in the plugin.
